Friends! (#20)
* rewrite crypto in the new protocol

* redesign daemon schema

* fix typo

* add story to sync friend request

* add story to public id response

* Public identifier (#18)

* add identifier.proto

* implement base58

* implement identifier

* Implement friend acking (#21)

* update message.proto with control message

* handle control message properly

* check for ack of control message

* Async friending schema (#22)

* configure server for async requests

* checkpoint: rewriting test files

* add friend workinggit ca add


* change api

* add sync test

* push without compile

* use enums instead of integer constants

* daemon.protoo changes

* async partitions and a basic test passes

* update proto. time for disaster to happen. Gooddddd luck

* update protos. We are good.

* update protos for Invitation Progres

* replace friend_request -> invitation on server

* Friends implementation (#23)

* change to system message

* it builds

* initial tests passing!!!

* revamp crypto

* fmt all

* add license notices

* more format fixes

* add asyncinvitation proto and additional check

* it compiles

+++ asphr/pir/fast_pir/fast_pir_config.hpp
@@ -55,7 +55,7 @@ static auto create_context_params() -> seal::EncryptionParameters {
   seal::EncryptionParameters params(seal::scheme_type::bfv);
   params.set_poly_modulus_degree(POLY_MODULUS_DEGREE);
   vector<seal::Modulus> coeff_modulus;
-  for (auto& prime : COEFF_MODULUS_FACTORIZATION) {
+  for (auto &prime : COEFF_MODULUS_FACTORIZATION) {
     coeff_modulus.push_back(seal::Modulus(prime));
   }
   params.set_coeff_modulus(coeff_modulus);
@@ -63,7 +63,7 @@ static auto create_context_params() -> seal::EncryptionParameters {
   return params;
 }
 
-constexpr size_t SEAL_DB_COLUMNS = CEIL_DIV(MESSAGE_SIZE_BITS, PLAIN_BITS);
+constexpr int SEAL_DB_COLUMNS = CEIL_DIV(MESSAGE_SIZE_BITS, PLAIN_BITS);
 
 // CLIENT_DB_ROWS is the number of rows that the client thinks is in the
 // database. this must be an upper bound on the actual database size. note that
@@ -77,4 +77,4 @@ constexpr size_t SEAL_DB_COLUMNS = CEIL_DIV(MESSAGE_SIZE_BITS, PLAIN_BITS);
 // is to be under the GRPC message size limit of 4 MB — to increase this, use
 // GRPC streaming instead. See
 // https://jbrandhorst.com/post/grpc-binary-blob-stream/.
-constexpr size_t CLIENT_DB_ROWS = 360'000;
+constexpr int CLIENT_DB_ROWS = 360'000;
